requirements.txt
pandas
scikit-learn
matplotlib
cassandra-driver